charts/requirements.txt
Flask==2.1.3 
# PyMySQL==1.0.2
python-dotenv==0.20.0 
flask-mysqldb==1.0.1